Add a new function to parse maps
This new function parses comma separated maps to keywords and doubles.
It was made to address issue #1489, to clarify the number of fields in
the multicomponent material model and similar plugins. In the current
format, values for various parameters such as density are assigned to
compositional fields by a list of values. However, which value in the
list goes to which compositional field is unclear. This new function
allows each value to be labeled with the id given to each field, and the
values can be listed in any order. The old functionality is retained,
and the new function can detect which format is provided in the
parameter file.