cmake: disallow in-source test runs
Running the whole testsuite in an in-source build has several problems.
Most importantly, the test build and source directory are not separate,
so the project can not be easily reset.
Rather than trying to fix this, disallow running the "all tests". Of
course running our single "quick test" is still okay.