Provide formulas for the material-statistics postprocessor.
Better be specific with formulas than rely on a reader's understanding of
a text description.

I'm not sure there is a reason why one would ever want to use anything
other than the usual arithmetic averaging for the density and mass,
given that these have physical interpretations, so I dropped that remark.